% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_ranges.R
\name{fill_ranges}
\alias{fill_ranges}
\title{Fills the gaps between ranges (where gap is defined as more than 1 day/second of difference).}
\usage{
fill_ranges(df, groups = NULL, start_var = NULL, end_var = NULL,
  dimension = "date", fmt = "\%Y-\%m-\%d", tz = "UTC",
  origin = "1970-01-01")
}
\arguments{
\item{df}{Your data frame}

\item{groups}{Grouping variables}

\item{start_var}{Start of the range}

\item{end_var}{End of the range}

\item{dimension}{Indicate whether your range includes only dates ('date') or also timestamp ('timestamp'). Defaults to 'date'}

\item{fmt}{The format of your date or timestamp field, defaults to YMD}

\item{tz}{Time zone, defaults to UTC}

\item{origin}{Origin for timestamp conversion, defaults to 1970-01-01}
}
\value{
Returns a data frame (if initial input data.table, then data.table) with missing ranges.
}
\description{
Columns that are not start/end dates or grouping variables are returned with NA as values for missing ranges.
}
\examples{
df <- data.frame(
group = c("a", "a", "b", "b", "b"),
start = c("2007-01-01", "2010-06-02", "2009-04-05", "2012-08-01", "2019-03-19"),
end = c("2008-02-05", "2013-04-05", "2009-06-03", "2013-02-17", "2021-04-21"),
cost = c(143, 144, 105, 153, 124)
)

fill_ranges(df, start_var = "start", end_var = "end", groups = c("group"))
}
